==> 01_data_preparation_and_aux_tables.ipynb <==
# 01 - Data Preparation and Auxiliary Tables

import pickle
import os
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
from scipy.signal import resample
from datetime import datetime

## Load data

#-- Importamos ds 1
sales = pd.read_csv('/Users/anapeyra/Desktop/NUCLIO/Entregables/TFM/stock-prediction-dsmarket/data_dsmarket/raw/item_sales.csv')

#-- Creamos lista de variables idenfiticadoras ('lista_vars_id') y variables de medidas ('lista_vars_dias')
lista_vars = sales.columns.values.tolist()
print(lista_vars)
lista_vars_id = lista_vars[:7]
print(lista_vars_id)
lista_vars_dias = lista_vars[7:]
print(lista_vars_dias)

# Utilizamos la función .melt
sales = sales.melt(id_vars=lista_vars_id, var_name="date", value_name= "units")

⚠️⚠️⚠️ La función **melt** se utiliza para convertir un DataFrame de ancho (wide) a un formato largo (long) en pandas. Permite deshacer la estructura de las columnas y apilarlas en una sola columna, mientras se conserva la información de las variables identificadoras.

En nuestro caso, al aplicar la función melt al DataFrame sales, especificamos las variables identificadoras en el parámetro **id_vars**, el nombre de la columna que almacenará las variables de fecha en el parámetro **var_name**, y el nombre de la columna que almacenará los valores de las unidades en el parámetro **value_name**.

sales.shape

sales.head()

sales.rename(columns={"date": "d"}, inplace = True)
sales.head()

sales.to_csv('/Users/anapeyra/Desktop/NUCLIO/Entregables/TFM/stock-prediction-dsmarket/data_dsmarket/modified/item_sales_mod.csv')

#-- Importamos ds 2
calendar = pd.read_csv('/Users/anapeyra/Desktop/NUCLIO/Entregables/TFM/stock-prediction-dsmarket/data_dsmarket/raw/daily_calendar_with_events.csv')
calendar.head()

calendar.shape

calendar.info()

calendar["event"].value_counts()

calendar["event"] = calendar["event"].fillna(0)

calendar["event"].isna().sum()

calendar["date"] = pd.to_datetime(calendar["date"], format = "%Y-%m-%d")
calendar["week"] = calendar["date"].dt.isocalendar().week.astype(str)
calendar["year"] = calendar["date"].dt.isocalendar().year.astype(str)
calendar["year_week"] = calendar["year"] + "-" + calendar["week"]

calendar.head()

calendar.to_csv('/Users/anapeyra/Desktop/NUCLIO/Entregables/TFM/stock-prediction-dsmarket/data_dsmarket/modified/daily_calendar_with_events_mod.csv')

#-- Importamos ds 3
item_prices = pd.read_csv('/Users/anapeyra/Desktop/NUCLIO/Entregables/TFM/stock-prediction-dsmarket/data_dsmarket/raw/item_prices.csv')
item_prices.head()

item_prices.shape

item_prices["yearweek2"] = item_prices['yearweek'].astype(str)
item_prices["year"] = item_prices['yearweek2'].str[:4]
item_prices["week"] = item_prices["yearweek2"].str[4:6]
item_prices["year_week"] = item_prices["year"] + "-" + item_prices["week"]
item_prices["id"] = item_prices["item"] + "_" + item_prices["store_code"]

item_prices.head()

item_prices.drop(["item", "category", "store_code", "yearweek", "yearweek2"], axis=1, inplace=True)
item_prices.head()

#-- Round 'sell_price' (two decimal places)
item_prices["sell_price"] = item_prices["sell_price"].round(2)
item_prices.head()

item_prices.to_csv('/Users/anapeyra/Desktop/NUCLIO/Entregables/TFM/stock-prediction-dsmarket/data_dsmarket/modified/item_prices_mod.csv')

## Merge data

#-- Mergeamos ds 1 & 2
sales_calendar = pd.merge(
    left = sales,
    right = calendar,
    how = 'left',
    on = ["d"])

sales_calendar.info()

sales_calendar.head()

sales_calendar.shape

sales_calendar.to_csv('/Users/anapeyra/Desktop/NUCLIO/Entregables/TFM/stock-prediction-dsmarket/data_dsmarket/modified/sales_calendar.csv')

### Auxiliary data ("year_week")
#### 'date_id': first 'date' for each 'year_week' and 'id' 

columns_to_keep = ["id", 
                   "date", 
                   "weekday", # verif always mismo day of week
                   "year_week",]
columns_to_drop = sales_calendar.columns.difference(columns_to_keep)
sales_calendar_1 = sales_calendar.drop(columns=columns_to_drop)
sales_calendar_1

date_id = sales_calendar_1.sort_values(by=['date', "year_week", "id"]).reset_index(drop=True)
date_id

date_id['date'] = pd.to_datetime(date_id['date'])

# Seleccionar la primera fecha para cada grupo "year_week", "id"
# De esta manera solo tenemos 1 día escogido (el 1º) apra cada year-week (p.e.: Monday)
date_id_2 = date_id.groupby(["year_week", "id"]).first()
date_id_2

date_id_2["weekday"].value_counts()

Observamos que hay datos de semana que empiezan en "Saturday", por lo que debe haber alguna semana incompleta.

date_id_2.drop(columns=["weekday"], inplace=True)

date_id_2.to_csv('/Users/anapeyra/Desktop/NUCLIO/Entregables/TFM/stock-prediction-dsmarket/data_dsmarket/modified/date_id.csv')

##### Verificamos si todas las semanas tienen el mismo recuento de registros

date_counts = date_id.groupby('year_week')['date'].nunique().reset_index()
date_counts.columns = ['year_week', 'day_count']
date_counts

date_counts['day_count'].value_counts()

date_counts[date_counts['day_count']==2]

Vemos que la primera semana (2011-4) solo tiene 2 días (deberíamos eliminar los registros más adelante)

#### "unit_id": ventas agrupadas por "id" y "year_week"

columns_to_keep = ["id", "units", "year_week"]
columns_to_drop = sales_calendar.columns.difference(columns_to_keep)
sales_calendar_2 = sales_calendar.drop(columns=columns_to_drop)

unit_id = sales_calendar_2.groupby(["year_week", "id"]).sum()
unit_id

unit_id.to_csv('/Users/anapeyra/Desktop/NUCLIO/Entregables/TFM/stock-prediction-dsmarket/data_dsmarket/modified/unit_id.csv')

## Merge ds (n1+n2) + n3:

item_prices.drop(columns=["year", "week"], inplace=True)
item_prices.head()

## Merge datasets (1+2) + 3:
sales_calendar_prices_full = pd.merge(
    left = sales_calendar,
    right = item_prices,
    how = 'left',
    on = ["id", "year_week"]
    )

sales_calendar_prices_full

sales_calendar_prices_full.shape

sales_calendar_prices_full.info()

sales_calendar_prices_full.to_csv('/Users/anapeyra/Desktop/NUCLIO/Entregables/TFM/stock-prediction-dsmarket/data_dsmarket/modified/sales_calendar_prices_full.csv')
