==> 02_table_merger.ipynb <==
import pickle
import os
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
from scipy.signal import resample
from datetime import datetime

unit_id = pd.read_csv('/Users/anapeyra/Desktop/NUCLIO/Entregables/TFM/stock-prediction-dsmarket/data_dsmarket/modified/unit_id.csv')
unit_id.head()

unit_id.shape

item_sales = pd.read_csv('/Users/anapeyra/Desktop/NUCLIO/Entregables/TFM/stock-prediction-dsmarket/data_dsmarket/raw/item_sales.csv')
item_sales.head()

cols = item_sales.columns.values.tolist()
cols_to_drop = cols[7:]
item_sales.drop(cols_to_drop, axis=1, inplace=True)

item_sales.head()

sales_calendar = pd.merge(
    left = unit_id,
    right = item_sales,
    how = "left",
    on = ["id"]
)

sales_calendar.head()

sales_calendar.shape

date_id = pd.read_csv('/Users/anapeyra/Desktop/NUCLIO/Entregables/TFM/stock-prediction-dsmarket/data_dsmarket/modified/date_id.csv')
date_id.head()

sales_calendar = pd.merge(
    left = sales_calendar,
    right = date_id,
    how = "left",
    on = ["id", "year_week" ]
)

sales_calendar.head()

sales_calendar.shape

item_prices = pd.read_csv('/Users/anapeyra/Desktop/NUCLIO/Entregables/TFM/stock-prediction-dsmarket/data_dsmarket/modified/item_prices_mod.csv')
item_prices.head()

item_prices.drop("Unnamed: 0", axis=1, inplace=True)
item_prices.head()

sales_calendar_prices = pd.merge(
    left = sales_calendar,
    right = item_prices,
    how = "left",
    on = ["id", "year_week" ]
)

sales_calendar_prices.head()

sales_calendar_prices.shape

sales_calendar_prices.to_csv('/Users/anapeyra/Desktop/NUCLIO/Entregables/TFM/stock-prediction-dsmarket/data_dsmarket/modified/sales_calendar_prices.csv')

## Other aux files:

gb_year = sales_calendar_prices.groupby(
    ["id", "item", "category", "department", "store", "region", "year"]
).sum(numeric_only=True)
gb_year.reset_index(inplace=True)

gb_year.head(10)

gb_year.drop("sell_price", axis=1, inplace=True)

gb_year.shape

gb_year.to_excel('/Users/anapeyra/Desktop/NUCLIO/Entregables/TFM/stock-prediction-dsmarket/data_dsmarket/modified/gb_year.xlsx')
