==> 06_general_processing_year_week.ipynb <==
import pickle
import os
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
from scipy.signal import resample
from datetime import datetime, timedelta

## Preprocessing

df = pd.read_csv('/Users/anapeyra/Desktop/NUCLIO/Entregables/TFM/stock-prediction-dsmarket/data_dsmarket/modified/sales_calendar_prices.csv')
df.drop('Unnamed: 0', axis=1, inplace=True)

df.head()

df.shape

df.info()

* Tenemos 'date' como object, habrá que pasarlo a fecha.
* Podemos pasar 'year' y 'week' a integers ✅, y luego también a un formato de fecha.
* Podemos generar también una columna de 'month' y de 'day', para acabar de tener todos los datos. ✅

# Transformar 'date' de object → datetime64[ns]
df["date"] = pd.to_datetime(df["date"])
df.info()

df[df.duplicated()]

df.isnull().sum()

df[df['sell_price'].isnull()].head()

* Para imputar los nulos de "sell_price" hay dos estrategias:

    1. Eliminar directamente las filas que contengan NaNs. ❌

# Porcentaje de datos que eliminaríamos:
3008732 / 8354260 * 100

# Igualemnte nos quedaríamos con:
8354260 - 3008732

df_without_nulls = df.dropna()
df_without_nulls.head()

df_without_nulls.isnull().sum()

df_without_nulls.shape

df_without_nulls.info()

2. Imputar los valores NaN en la columna 'sell_price' con los valores más recientes de la misma colmna, teniendo en cuenta el mismo 'id'. Vamosa  usar el método **bfilll()** junto con **groupby()** para realizar el relleno hacia atrás por grupos. Primero vamos a asegurarnos que los datos estén ordenado antes de realizar el bakfill. ✅

df_filled = df.sort_values(['id','date'])
df_filled.head()

df_filled['sell_price'] = df_filled.groupby(['item'])['sell_price'].fillna(method='ffill').fillna(method='bfill')
df_filled.head()

df_filled["week"] = df_filled["date"].dt.isocalendar().week.astype(str)
df_filled["year"] = df_filled["date"].dt.isocalendar().year.astype(str)
df_filled.head()

# Usamos el df_filled
df_final = df_filled.copy()
df_final.head()

df_final['year'] = df_final["year"].apply(int)
df_final["month_int"] = df_final["date"].dt.month.astype(str)
df_final["month"] = df_final["date"].dt.strftime("%B")
df_final['week'] = df_final["week"].apply(int)
df_final["day"] = df_final["date"].dt.day.astype(str)

df_final.head()

* Vamos a eliminar los registros de la 1ª semana, ya que vimos (en "1_aux_db") que estaban incompletos.

df_final = df_final[df_final["year_week"] != "2011-4"]
df_final

## Feature engineering - General features

df_final['revenue'] = df_final['units'] * df_final['sell_price']
df_final.head()

### Campaña de eventos (extendidos)

calendar = pd.read_csv('/Users/anapeyra/Desktop/NUCLIO/Entregables/TFM/stock-prediction-dsmarket/data_dsmarket/modified/daily_calendar_with_events_mod.csv')
calendar.head()

calendar.info()

# Transformar 'date' de object -> datetime64[ns]
calendar["date"] = pd.to_datetime(calendar["date"])

calendar.drop(columns = ['Unnamed: 0'], inplace=True)
calendar.head()

calendar["event"].value_counts()

# Vamos a generar una nueva columna para cada uno de los eventos, apra tratarlos de manera inidividual
calendar["event_Ramadan"] = np.where(calendar["event"] == "Ramadan starts", 1, 0)
calendar["event_SuperBowl"] = np.where(calendar["event"] == "SuperBowl", 1, 0)
calendar["event_Thanksgiving"] = np.where(calendar["event"] == "Thanksgiving", 1, 0)
calendar["event_NewYear"] = np.where(calendar["event"] == "NewYear", 1, 0)
calendar["event_Easter"] = np.where(calendar["event"] == "Easter", 1, 0)

calendar.info()

* Vamos a tener en cuenta que la campaña de ventas por los eventos de (SuperBowl, Thanksgiving, NewYear y Easter) empieza 1 semana antes para cada uno de los eventos. ✅

def extend_events_previous_days(df, columns, previous_days):
    # Asegúrate de que los datos estén ordenados por fecha
    df = df.sort_values(by='date')

    # Crea un DataFrame temporal que marque los eventos y los días anteriores como eventos
    df_temp = df.copy()

    for column in columns:
        # Crea una columna extendida para el evento actual
        df_temp[column + '_extended'] = df_temp[column]

        # Extiende los eventos hacia los días anteriores
        for i in range(1, previous_days):
            shifted = df[column].shift(i).fillna(0).astype(int)
            df_temp[column + '_extended'] = df_temp[column + '_extended'] | shifted

        # Reemplaza la columna original con la columna extendida
        df[column] = df_temp[column + '_extended']

    # Limpia
    del df_temp

    return df

calendar["event_SuperBowl"].value_counts()

calendar = extend_events_previous_days(df = calendar, 
                                       columns =["event_SuperBowl",                                                
                                                 "event_Thanksgiving", 
                                                 "event_NewYear", 
                                                 "event_Easter"],
                                       previous_days = 7)

calendar["event_SuperBowl"].value_counts()

# Verificamos que realmente se han añadido 7 días más a cada evento:
6*7

* En el caso del Ramadán vamos a tener en cuenta que dura un mes, así que vamos a tener en cuenta el inicio +30 días después de que empiece.

def extend_events_posterior_days(df, columns, posterior_days):
    # Asegúrate de que los datos estén ordenados por fecha
    df = df.sort_values(by='date')

    # Crea un DataFrame temporal que marque los eventos y los días posteriores como eventos
    df_temp = df.copy()

    for column in columns:
        # Crea una columna extendida para el evento actual
        df_temp[column + '_extended'] = df_temp[column]

        # Extiende los eventos hacia los días posteriores
        for i in range(1, posterior_days):
            shifted = df[column].shift(-i).fillna(0).astype(int)
            df_temp[column + '_extended'] = df_temp[column + '_extended'] | shifted

        # Reemplaza la columna original con la columna extendida
        df[column] = df_temp[column + '_extended']

    # Clean
    del df_temp

    return df

calendar["event_Ramadan"].value_counts()

calendar = extend_events_posterior_days(df = calendar, 
                                        columns = ["event_Ramadan"],
                                        posterior_days = 30)

calendar["event_Ramadan"].value_counts()

#-- Verify +7 por evento
5*30

#-- Dummy generate (all eventos ext)
calendar['event_any'] = np.where(calendar[['event_SuperBowl', 
                                           'event_Thanksgiving', 
                                           'event_Easter',
                                           'event_NewYear',
                                           'event_Ramadan']].any(axis=1), 1, 0)
calendar['event_any'].value_counts()

calendar.drop(columns="event", inplace=True)

calendar.head()

calendar.info()

calendar.to_csv('/Users/anapeyra/Desktop/NUCLIO/Entregables/TFM/stock-prediction-dsmarket/data_dsmarket/interim/daily_calendar_with_events_mod_extended.csv')

calendar.drop(columns=["d", 'week', 'year', 'year_week'], inplace=True)

calendar.info()

df_final = pd.merge(
    left = df_final,
    right = calendar,
    how = "left",
    on = ["date"]
)

df_final.head()

df_final.info()

#-- Orden columnas (our choice)
column_order = ['date',
                'year',
                'month',
                'month_int',
                'week',
                'year_week',
                'day',
                'weekday',
                'weekday_int',
                'event_any',
                'event_SuperBowl',
                'event_Thanksgiving',
                'event_Ramadan',
                'event_NewYear',
                'event_Easter', 
                'id',
                'item',
                'category',
                'department',
                'store', 
                'store_code', 
                'region',
                'units',
                'sell_price',
                'revenue']

df_final = df_final.reindex(columns=column_order)
df_final.head()

### Otras campañas:

#-- Create column ventas verano
df_final['summer_sales'] = np.where(df_final["month"].isin(["June","July", "August"]), df_final.units, 0)

#-- Create column ventas Navidad
df_final['christmas_sales'] = np.where(df_final["month"].isin(["November","December"]), df_final.units, 0)

df_final.head()

#-- Create column con ventas en eventos
df_final["event_sales"] = np.where(df_final["event_any"] == 1, df_final.units, 0)

df_final["event_SuperBowl_sales"] = np.where(df_final["event_SuperBowl"] == 1, df_final.units, 0)
df_final["event_Thanksgiving_sales"] = np.where(df_final["event_Thanksgiving"] == 1, df_final.units, 0)
df_final["event_Ramadan_sales"] = np.where(df_final["event_Ramadan"] == 1, df_final.units, 0)
df_final["event_NewYear_sales"] = np.where(df_final["event_NewYear"] == 1, df_final.units, 0)
df_final["event_Easter_sales"] = np.where(df_final["event_Easter"] == 1, df_final.units, 0)

df_final.info()

### Ventas por año

df_final["sales_2011"] = np.where((df_final.year == 2011), df_final.units, 0) # Incomplete
df_final["sales_2012"] = np.where((df_final.year == 2012), df_final.units, 0)
df_final["sales_2013"] = np.where((df_final.year == 2013), df_final.units, 0)
df_final["sales_2014"] = np.where((df_final.year == 2014), df_final.units, 0)
df_final["sales_2015"] = np.where((df_final.year == 2015), df_final.units, 0)
df_final["sales_2016"] = np.where((df_final.year == 2016), df_final.units, 0) # Incomplete

df_final.sample(5)

#-- Obtener fecha máx. de 2016
max_date = df_final['date'].max()
max_date

#-- Create selección para fechas hasta el 18 de abril de 2015
df_final["sales_2015_rel"] = np.where(((df_final.year == 2015) & (df_final.date <= '2015-04-18')), df_final.units, 0)

df_final.info()

df_final.to_csv('/Users/anapeyra/Desktop/NUCLIO/Entregables/TFM/stock-prediction-dsmarket/data_dsmarket/interim/df_forecast_1.csv')
