==> 07_general_preprocessing_days.ipynb <==
import pickle
import os
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
from scipy.signal import resample
from datetime import datetime

### Merge datasets 1_mod + 2_mod:

sales = pd.read_csv('/Users/anapeyra/Desktop/NUCLIO/Entregables/TFM/stock-prediction-dsmarket/data_dsmarket/modified/item_sales_mod.csv')
del(sales["Unnamed: 0"])
sales.head()


calendar = pd.read_csv('/Users/anapeyra/Desktop/NUCLIO/Entregables/TFM/stock-prediction-dsmarket/data_dsmarket/modified/daily_calendar_with_events_mod.csv')

del(calendar["Unnamed: 0"])
calendar.head()

## Merge datasets 1_mod + 2_mod:
df = pd.merge(
    left = sales,
    right = calendar,
    how = 'left',
    on = ["d"]
    )
# Archivo similar a: "sales_calendar.csv"
df

## Preprocessing

df.shape

df.info()

## Feature engineering - Part 2 - Using all data ("days")

### Item clustering

#### Week/Weekend_sales

# Creamos una columna que tenga las ventas de los fines de semana:
df["weekend_sales"] = np.where(df["weekday"].isin(["Saturday", "Sunday"]), df.units, 0)

# Creamos una columna que tenga las ventas durante la semana:
df["week_sales"] = np.where(df["weekday"].isin(["Monday", "Tuesday", "Wednesday", "Thursday", "Friday"]), df.units, 0)

df

df_cluster_item_extra = df.groupby("item")[["week_sales", "weekend_sales"]].sum()
df_cluster_item_extra = df_cluster_item_extra.reset_index()
df_cluster_item_extra

#### Encontrar las fechas de la primera y última compra para cada artículo:

# Todos los valores
print(len(df))

# Valores con ventas
df_sales = df[df['units'] != 0]
print(len(df_sales))

# Valores sin ventas
df_nans = df[df['units'] == 0]
print(len(df_nans))

print()

#  Se agrupa el DataFrame por la columna "item" y se obtiene la primera fecha con ventas ("df_final_sales")
item_first_last_sale = df_sales.groupby(["item"]).agg({'date': [np.min, np.max]})
item_first_last_sale = item_first_last_sale.reset_index()

# Resetear el nombre de las columnas a un rango de valores numericos
item_first_last_sale.columns = range(item_first_last_sale.shape[1])
# Crear diccionario para mapeo de nombres de columna
lista_1 = item_first_last_sale.columns.tolist()
lista_2 = ["item", "item_first_sale", "item_last_sale"]
dict_dates = dict(zip(lista_1, lista_2))

item_first_last_sale.rename(columns=dict_dates, inplace=True)
item_first_last_sale

#### Calcular el tiempo de exposición:

item_first_last_sale.info()

item_first_last_sale["item_first_sale"] = pd.to_datetime(item_first_last_sale["item_first_sale"])
item_first_last_sale["item_last_sale"] = pd.to_datetime(item_first_last_sale["item_last_sale"])

item_first_last_sale.info()

item_first_last_sale['item_exhibition_days'] = (item_first_last_sale['item_last_sale'] - item_first_last_sale['item_first_sale']).dt.days
item_first_last_sale.head()

df_cluster_item_extra = pd.merge(
    left = df_cluster_item_extra,
    right = item_first_last_sale,
    how = 'left',
    on = ["item"])

df_cluster_item_extra.head()

#### Días de compras al año

# Agrupa por ítem y calcula el número de días únicos con compras para cada ítem
num_days_with_purchases_by_item = df_sales.groupby('item')['date'].nunique() #se trata de un array

# Convierte la serie en un DataFrame
df_days_with_purchases = num_days_with_purchases_by_item.reset_index()
df_days_with_purchases.columns = ['item', 'days_with_purchases']

# Limpia
del(num_days_with_purchases_by_item)

df_days_with_purchases

df_cluster_item_extra = pd.merge(
    left = df_cluster_item_extra,
    right = df_days_with_purchases,
    how = 'left',
    on = ["item"])

df_cluster_item_extra.head()

#### Ratio (días de compras/tiempo de exposición)

df_cluster_item_extra["purchase_to_exhib_ratio"] = df_cluster_item_extra["days_with_purchases"] / df_cluster_item_extra["item_exhibition_days"]
df_cluster_item_extra.head()

df_cluster_item_extra.to_csv('/Users/anapeyra/Desktop/NUCLIO/Entregables/TFM/stock-prediction-dsmarket/data_dsmarket/task_2_clustering/df_cluster_item_extra.csv', index=False)
